Move clients types to pkg/clients/
